Create a matrix of predictions for a caret_stack object.

Create a matrix of predictions for a caret_stack object.

Usage

# S3 method for class 'caret_stack'
predict(object, data_list, drop_redundant_class = TRUE, ...)

Arguments

object

A caret_stack object

data_list

A list of datasets to predict on, with each dataset matching the corresponding model in caret_list.

drop_redundant_class

A boolean controlling whether to exclude the first class from prediction output. Default is TRUE.

...

Additional arguments to pass to caret::predict.

Value

A data.table::data.table of predictions for base and ensemble models.

Examples

# Load example data and pre-trained caret_stack object
data(heart_failure_datasets)
data(heart_failure_stack)

# List of datasets to predict on
data_list <- heart_failure_datasets[c("cells", "holter", "mrna", "proteins")]

predict(heart_failure_stack, data_list)
#>         cells     holter       mrna     proteins   ensemble
#>         <num>      <num>      <num>        <num>      <num>
#>  1: 0.4049099 0.35599839 0.60550146 9.999201e-01 0.93426721
#>  2: 0.3189746 0.18674512 0.64169176 9.494278e-01 0.86220391
#>  3: 0.3410794 0.22072744 0.55084082 9.999484e-01 0.85274768
#>  4: 0.1672318 0.27355821 0.55445275 9.241026e-01 0.86516975
#>  5: 0.4327473 0.30317731 0.77407516 9.443000e-01 0.94968885
#>  6: 0.2613006 0.35501503 0.07158567 2.411954e-02 0.14523570
#>  7: 0.2351919 0.67202485 0.36258531 9.238138e-01 0.95621505
#>  8: 0.1876449 0.21599038 0.56700482 9.737941e-01 0.85026060
#>  9: 0.2727293 0.26661519 0.66887927 9.641809e-01 0.91511901
#> 10: 0.3556961 0.28240732 0.51360373 9.594888e-01 0.86073449
#> 11: 0.4227678 0.33519316 0.70047505 9.637260e-01 0.94519366
#> 12: 0.1408136 0.24019401 0.20400964 9.772814e-01 0.60009980
#> 13: 0.2376761 0.62918747 0.44666837 9.588456e-01 0.96404426
#> 14: 0.2539488 0.25509892 0.65634450 9.377611e-01 0.90097363
#> 15: 0.1647435 0.20081932 0.02877676 1.534899e-05 0.05876004
#> 16: 0.3270794 0.12234533 0.29117016 3.291748e-03 0.10871622
#> 17: 0.2443443 0.14278646 0.14348725 3.418846e-04 0.06898305
#> 18: 0.2123398 0.15902511 0.16666602 2.338829e-03 0.08142600
#> 19: 0.1474506 0.18339779 0.04581124 1.076370e-04 0.05774057
#> 20: 0.1686462 0.03440070 0.13062402 5.756124e-06 0.03923285
#> 21: 0.3769841 0.17877535 0.19292230 2.669167e-05 0.09762476
#> 22: 0.2822315 0.10292835 0.19533069 6.137831e-03 0.07042717
#> 23: 0.2570507 0.10233610 0.09137107 1.685702e-02 0.04850670
#> 24: 0.1580994 0.20683761 0.11113430 7.702148e-03 0.08360380
#> 25: 0.2609020 0.15235890 0.15134707 8.614690e-04 0.07435724
#> 26: 0.1062625 0.27661951 0.07277311 6.630702e-03 0.09962520
#> 27: 0.1666454 0.18269704 0.08590046 1.172674e-02 0.06865787
#> 28: 0.1489136 0.19170318 0.11522146 2.932888e-06 0.07786464
#> 29: 0.4150281 0.21445948 0.23750161 2.840306e-02 0.14190199
#> 30: 0.2339490 0.11192896 0.23205740 9.409089e-02 0.10134831
#> 31: 0.1583736 0.32434875 0.04749799 7.798561e-05 0.11114473
#> 32: 0.2083228 0.36332906 0.08683575 1.360201e-06 0.15114461
#> 33: 0.1596969 0.10165365 0.11842796 6.019713e-03 0.05234347
#> 34: 0.1478155 0.18210726 0.12506201 2.571664e-05 0.07725934
#> 35: 0.1373942 0.30894452 0.06850441 9.469738e-03 0.11405989
#> 36: 0.1520908 0.17948732 0.07652460 7.053547e-04 0.06374609
#> 37: 0.2098226 0.24091567 0.12678188 2.641588e-02 0.10745617
#> 38: 0.2806649 0.20478526 0.26759854 2.845335e-02 0.15098638
#> 39: 0.1553361 0.12179211 0.16851209 3.134787e-02 0.07345920
#> 40: 0.2342897 0.37628812 0.06584982 3.056551e-02 0.15790613
#> 41: 0.1524704 0.26659514 0.15898760 9.433155e-04 0.12807548
#> 42: 0.2025062 0.08928739 0.06198357 2.125470e-03 0.03938148
#> 43: 0.1831144 0.22524699 0.06791555 2.138992e-02 0.07987531
#> 44: 0.3337446 0.27084279 0.21955745 3.425680e-02 0.17143258
#> 45: 0.1127272 0.28892694 0.09622044 4.672691e-02 0.12434111
#> 46: 0.2128502 0.26930293 0.20773046 3.796104e-03 0.15423213
#> 47: 0.2031294 0.10648009 0.03331503 1.339742e-03 0.03822613
#> 48: 0.1214292 0.12111786 0.06296663 1.712670e-05 0.04583917
#> 49: 0.1245156 0.26095248 0.13678298 1.445131e-02 0.11876158
#> 50: 0.1519427 0.26223041 0.08741950 2.399118e-02 0.10209626
#> 51: 0.2790865 0.14203346 0.26751638 1.151787e-09 0.10831597
#> 52: 0.1416344 0.13534511 0.09102516 3.251749e-05 0.05460515
#> 53: 0.3376106 0.13927328 0.25292449 2.785787e-05 0.10151278
#> 54: 0.2118272 0.24031938 0.08924377 1.627186e-04 0.08849400
#> 55: 0.1444654 0.22452562 0.08191330 1.054696e-02 0.08191428
#> 56: 0.1671825 0.21010047 0.14459207 2.725201e-02 0.09994302
#> 57: 0.2362149 0.09313817 0.12750903 6.055067e-04 0.05142020
#> 58: 0.1354322 0.09452217 0.05087094 1.449896e-06 0.03850523
#>         cells     holter       mrna     proteins   ensemble
#>         <num>      <num>      <num>        <num>      <num>